+++ CHANGELOG.md
@@ -6,6 +6,12 @@ Note: once you update a project on that uses this template, be sure to copy this
 
 # Released 
 
+## v2.1.3 - 05/29/2025
+
+Improves the overall layout app functionality.
+
+- Updated layoutApp with more example options with title and custom width.
+
 ## v2.1.2 - 05/22/2025
 
 Adds examples of datatables atlas component and composable usage with inertia response.